TO-DO LIST APP.py
CODSOFT: Python To-Do List App ( I completed my project task one   for an internship in CODSOFT)
This repository contains a simple To-Do List application implemented in Python using the 'tkinter' library for the graphical user interface. The application allows users to manage their tasks by adding, deleting, saving, and loading tasks from a file.
Features:
Add Tasks: Users can add new tasks to the To-Do List by entering the task description in the provided entry field and clicking the "Add Task" button.

Delete Tasks: Users can remove tasks from the To-Do List by selecting a task from the list and clicking the "Delete Task" button.

Save Tasks: The application allows users to save the current list of tasks to a file named "tasks.dat" using the "Save Tasks" button.

Load Tasks: Users can load previously saved tasks from the "tasks.dat" file into the To-Do List using the "Load Tasks" button.

Technologies Used:
Python
Tkinter (for GUI)

Code Overview:
Initialization: The main window of the application is created using tkinter.Tk() with a title "To-Do List App" and a pink background.

Header Label: A header label "To-Do List App" is created and displayed at the top of the main window.

Toolbar: A toolbar is created at the top of the main window to hold buttons for adding, deleting, saving, and loading tasks.

Task List: A Listbox widget is used to display the list of tasks added by the user.

Task Entry Field: An entry field is provided for users to type new tasks.

Add Task Function: A function addTask() is defined to add tasks to the Listbox when the "Add Task" button is clicked. If no task is entered, a warning message is displayed.

Delete Task Function: A function deleteTask() is defined to delete selected tasks from the Listbox when the "Delete Task" button is clicked. If no task is selected, a warning message is displayed.

Save Tasks Function: A function saveTasks() is defined to save the current list of tasks to a file "tasks.dat" when the "Save Tasks" button is clicked.

Load Tasks Function: A function loadTasks() is defined to load previously saved tasks from the "tasks.dat" file into the Listbox when the "Load Tasks" button is clicked. If no tasks are saved, a warning message is displayed.

Main Loop: The root.mainloop() function is called to start the main event loop and display the GUI, allowing users to interact with the To-Do List App.